requirements.txt
biopython
matplotlib
opencv-python